Automate Ghidra recovery convergence and export validation

+++ decomp/ST.exe/functions/0041D590/decomp.c
@@ -10,8 +10,8 @@ void __thiscall TLOBaseTy::sub_0041D590(TLOBaseTy *this)
   if ((g_visibleClass_00802A88 != (VisibleClassTy *)0x0) && ((this->field_01D1 & 1) == 0)) {
     VisibleClassTy::VisHoleCreate
               (g_visibleClass_00802A88,(int)this->field_005B,(int)this->field_005D,
-               (undefined *)(int)this->field_005F,this->field_0024,(undefined *)this->field_0101,
-               0x19);
+               (undefined *)(int)this->field_005F,(uint)this->field_0024,
+               (undefined *)this->field_0101,0x19);
   }
   return;
 }
