+++ decomp/ST.exe/functions/004DF9F0/decomp.c
@@ -11,7 +11,7 @@ undefined4 __thiscall TLOBaseTy::sub_004DF9F0(TLOBaseTy *this)
   uint index;
   TLOBaseTy *local_8;
 
-  array = g_playerRuntime[this->field_0024].field2165_0x9d2;
+  array = g_playerRuntime[(int)this->field_0024].field2165_0x9d2;
   if (array != (DArrayTy *)0x0) {
     index = 0;
     local_8 = this;
@@ -19,16 +19,16 @@ undefined4 __thiscall TLOBaseTy::sub_004DF9F0(TLOBaseTy *this)
       do {
         DArrayGetElement(array,index,&local_8);
         if (local_8 == this) {
-          FUN_006b0c70(g_playerRuntime[this->field_0024].field2165_0x9d2,index);
+          FUN_006b0c70(g_playerRuntime[(int)this->field_0024].field2165_0x9d2,index);
           break;
         }
         index = index + 1;
-        array = g_playerRuntime[this->field_0024].field2165_0x9d2;
+        array = g_playerRuntime[(int)this->field_0024].field2165_0x9d2;
       } while ((int)index < (int)array->count);
     }
-    if ((g_playerRuntime[this->field_0024].field2165_0x9d2)->count == 0) {
-      DArrayDestroy(g_playerRuntime[this->field_0024].field2165_0x9d2);
-      g_playerRuntime[this->field_0024].field2165_0x9d2 = (DArrayTy *)0x0;
+    if ((g_playerRuntime[(int)this->field_0024].field2165_0x9d2)->count == 0) {
+      DArrayDestroy(g_playerRuntime[(int)this->field_0024].field2165_0x9d2);
+      g_playerRuntime[(int)this->field_0024].field2165_0x9d2 = (DArrayTy *)0x0;
     }
   }
   return 0;
